mepo: add v1.52.0
Update mepo package to have v1.52.0. I'm working on a geosgcm package and the blobless clone ability of mepo 1.52 is nice to have.

Also, I'm adding @pchakraborty as a maintainer as, well, he is pretty much mepo now. Indeed, soon we hope to have mepo 2.0 out which will be a pip package which means this package will be changed a lot!